Do not attempt to provision initial resources when restarting
